aulas/modulo4/parte1(1)/src/modules/appointments/services/CreateAppointmentService.spec.ts
import 'reflect-metadata';

import AppError from '@shared/errors/AppError';
import CreateAppointmentService from './CreateAppointmentService';
import FakeAppointmentsRepository from '../repositories/fakes/FakeAppointmentsRepository';

describe('CreateAppointment', () => {
    it('should be able to create a new appointment', async () => {
        const fakeAppoinmentsRepository = new FakeAppointmentsRepository();
        const createAppointmentService = new CreateAppointmentService(
            fakeAppoinmentsRepository,
        );
        const appointment = await createAppointmentService.execute({
            date: new Date(),
            provider_id: '123123123',
        });

        expect(appointment).toHaveProperty('id');
        expect(appointment.provider_id).toBe('123123123');
    });

    it('should not be able to create two appointments on the same time', async () => {
        const fakeAppoinmentsRepository = new FakeAppointmentsRepository();
        const createAppointmentService = new CreateAppointmentService(
            fakeAppoinmentsRepository,
        );

        const appointmentDate = new Date(2020, 5, 10, 16);

        await createAppointmentService.execute({
            date: appointmentDate,
            provider_id: '123123123',
        });

        expect(
            createAppointmentService.execute({
                date: appointmentDate,
                provider_id: '123123123',
            }),
        ).rejects.toBeInstanceOf(AppError);
    });
});
